Moved out some legacy code into old-tools
